+++ src/styles/HUD.css
@@ -1,0 +1,91 @@
+.ui-overlay {
+    position: absolute;
+    top: 0;
+    left: 0;
+    right: 0;
+    height: 48px;
+    padding: 0 18px;
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    background: rgba(18, 19, 20, 0.2);
+    backdrop-filter: blur(12px);
+    -webkit-backdrop-filter: blur(12px);
+    border-bottom: 1px solid rgba(255, 255, 255, 0.12);
+    color: #f3f4f6;
+    z-index: 10;
+    box-shadow: 0 10px 24px rgba(0, 0, 0, 0.28);
+}
+
+.ui-overlay .item-count {
+    margin-top: 0;
+    font-weight: 600;
+    color: rgba(243, 244, 246, 0.74);
+    font-size: 11px;
+    letter-spacing: 0.08em;
+    text-transform: uppercase;
+}
+
+.toolbar {
+    display: flex;
+    gap: 12px;
+    align-items: center;
+    justify-content: flex-end;
+}
+
+.hud-btn {
+    display: inline-flex;
+    align-items: center;
+    gap: 6px;
+    padding: 0 12px;
+    height: 28px;
+    background: rgba(31, 41, 55, 0.42);
+    border: 1px solid rgba(255, 255, 255, 0.1);
+    border-radius: 4px;
+    color: #e5e7eb;
+    font-family: inherit;
+    font-size: 11px;
+    font-weight: 600;
+    letter-spacing: 0.08em;
+    text-transform: uppercase;
+    cursor: pointer;
+    box-shadow: none;
+    transition:
+        background 0.15s,
+        border-color 0.15s,
+        color 0.15s;
+}
+
+.hud-icon-btn {
+    width: 28px;
+    min-width: 28px;
+    justify-content: center;
+    padding: 0;
+}
+
+.hud-title {
+    font-size: 12px;
+    font-weight: 700;
+    letter-spacing: 0.14em;
+    color: #f8fafc;
+    text-transform: none;
+}
+
+.hud-btn svg {
+    opacity: 0.7;
+    flex-shrink: 0;
+}
+
+.hud-btn:hover {
+    background: rgba(17, 24, 39, 0.65);
+    border-color: rgba(229, 231, 235, 0.24);
+    color: #f9fafb;
+}
+
+.hud-btn:hover svg {
+    opacity: 1;
+}
+
+.hud-btn:active {
+    background: rgba(17, 24, 39, 0.82);
+}
