+++ src/styles/SettingsModal.css
@@ -1,0 +1,198 @@
+.settings-modal-backdrop {
+    position: fixed;
+    inset: 0;
+    z-index: 40;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    padding: 48px;
+    background: rgba(3, 7, 18, 0.58);
+    backdrop-filter: blur(5px);
+    -webkit-backdrop-filter: blur(5px);
+}
+
+.settings-modal {
+    width: min(920px, calc(100vw - 64px));
+    height: min(620px, calc(100vh - 96px));
+    min-height: 420px;
+    display: flex;
+    flex-direction: column;
+    overflow: hidden;
+    color: #f8fafc;
+    background: rgba(16, 18, 22, 0.96);
+    border: 1px solid rgba(255, 255, 255, 0.14);
+    border-radius: 8px;
+    box-shadow: 0 24px 80px rgba(0, 0, 0, 0.52);
+    outline: none;
+}
+
+.settings-modal-header {
+    height: 58px;
+    padding: 0 18px 0 24px;
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    border-bottom: 1px solid rgba(255, 255, 255, 0.1);
+}
+
+.settings-modal-header h2 {
+    margin: 0;
+    font-size: 15px;
+    line-height: 1;
+    font-weight: 700;
+    letter-spacing: 0;
+    text-transform: uppercase;
+}
+
+.settings-close-btn {
+    width: 32px;
+    height: 32px;
+    padding: 0;
+    display: inline-flex;
+    align-items: center;
+    justify-content: center;
+    color: #d1d5db;
+    background: rgba(255, 255, 255, 0.04);
+    border: 1px solid rgba(255, 255, 255, 0.1);
+    border-radius: 6px;
+    box-shadow: none;
+    line-height: 1;
+    cursor: pointer;
+}
+
+.settings-close-btn svg {
+    display: block;
+}
+
+.settings-close-btn:hover {
+    color: #ffffff;
+    background: rgba(255, 255, 255, 0.1);
+    border-color: rgba(255, 255, 255, 0.2);
+}
+
+.settings-modal-body {
+    min-height: 0;
+    flex: 1;
+    display: flex;
+}
+
+.settings-sidebar {
+    width: 220px;
+    padding: 18px 12px;
+    display: flex;
+    flex-direction: column;
+    background: rgba(0, 0, 0, 0.16);
+    border-right: 1px solid rgba(255, 255, 255, 0.1);
+}
+
+.settings-nav {
+    display: grid;
+    gap: 6px;
+}
+
+.settings-nav-item {
+    width: 100%;
+    height: 38px;
+    padding: 0 12px;
+    display: flex;
+    align-items: center;
+    color: rgba(243, 244, 246, 0.72);
+    background: transparent;
+    border: 1px solid transparent;
+    border-radius: 6px;
+    box-shadow: none;
+    font-family: inherit;
+    font-size: 13px;
+    font-weight: 600;
+    line-height: 1;
+    text-align: left;
+    cursor: pointer;
+}
+
+.settings-nav-item:hover {
+    color: #ffffff;
+    background: rgba(255, 255, 255, 0.06);
+    border-color: rgba(255, 255, 255, 0.08);
+}
+
+.settings-nav-item.active,
+.settings-nav-item[aria-pressed="true"] {
+    color: #ffffff;
+    background: rgba(20, 184, 166, 0.18);
+    border-color: rgba(20, 184, 166, 0.36);
+}
+
+.settings-version {
+    margin-top: auto;
+    padding: 14px 12px 0;
+    color: rgba(243, 244, 246, 0.54);
+    border-top: 1px solid rgba(255, 255, 255, 0.08);
+    font-size: 11px;
+    font-weight: 600;
+    letter-spacing: 0;
+    line-height: 1.2;
+    text-transform: uppercase;
+}
+
+.settings-panel {
+    flex: 1;
+    min-width: 0;
+    padding: 26px 30px;
+}
+
+.settings-panel h3 {
+    margin: 0;
+    color: rgba(248, 250, 252, 0.88);
+    font-size: 18px;
+    font-weight: 700;
+    line-height: 1.2;
+}
+
+.settings-toggle-row {
+    margin-top: 24px;
+    min-height: 48px;
+    padding: 0 16px;
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    gap: 16px;
+    color: rgba(248, 250, 252, 0.88);
+    background: rgba(255, 255, 255, 0.04);
+    border: 1px solid rgba(255, 255, 255, 0.08);
+    border-radius: 8px;
+    font-size: 13px;
+    font-weight: 600;
+    cursor: pointer;
+}
+
+.settings-toggle-row input {
+    width: 18px;
+    height: 18px;
+    margin: 0;
+    accent-color: #14b8a6;
+    box-shadow: none;
+    cursor: pointer;
+}
+
+@media (max-width: 700px) {
+    .settings-modal-backdrop {
+        padding: 24px;
+    }
+    .settings-modal {
+        width: calc(100vw - 32px);
+        height: calc(100vh - 48px);
+        min-height: 0;
+    }
+    .settings-modal-body {
+        flex-direction: column;
+    }
+    .settings-sidebar {
+        width: auto;
+        min-height: 180px;
+        border-right: none;
+        border-bottom: 1px solid rgba(255, 255, 255, 0.1);
+    }
+    .settings-panel {
+        padding: 22px;
+    }
+}
