+++ src/styles/VideoPlayer.css
@@ -1,0 +1,279 @@
+.video-lod-thumbnail {
+    filter: saturate(0.9) contrast(0.95);
+}
+
+.video-lod-proxy {
+    position: absolute;
+    inset: 0;
+    box-sizing: border-box;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    border: 2px solid rgba(94, 234, 212, 0.7);
+    background:
+        linear-gradient(
+            135deg,
+            rgba(20, 184, 166, 0.18),
+            rgba(15, 23, 42, 0.72)
+        ),
+        rgba(15, 23, 42, 0.8);
+    pointer-events: none;
+}
+
+.video-load-proxy {
+    width: 100%;
+    height: 100%;
+    padding: 0;
+    flex-direction: column;
+    gap: 14px;
+    color: rgba(226, 232, 240, 0.92);
+    font: inherit;
+    cursor: pointer;
+    pointer-events: auto;
+}
+
+.video-load-thumbnail {
+    opacity: 0.7;
+}
+
+.video-load-label {
+    font-size: 16px;
+    font-weight: 700;
+    z-index: 1;
+}
+
+.video-load-requested,
+.video-playback-error {
+    pointer-events: auto;
+}
+
+.video-timeline {
+    position: absolute;
+    left: 12px;
+    right: 12px;
+    bottom: 12px;
+    display: grid;
+    grid-template-columns: auto minmax(0, 1fr) auto auto;
+    gap: 10px;
+    align-items: center;
+    padding: 8px 10px;
+    border: 1px solid rgba(255, 255, 255, 0.16);
+    border-radius: 8px;
+    background: rgba(15, 23, 42, 0.72);
+    backdrop-filter: blur(10px);
+    -webkit-backdrop-filter: blur(10px);
+    color: rgba(248, 250, 252, 0.94);
+    opacity: 0;
+    pointer-events: none;
+    transform: translateY(6px);
+    transition:
+        opacity 0.16s ease,
+        transform 0.16s ease;
+    z-index: 4;
+}
+
+.media-item:hover .video-timeline,
+.video-timeline:focus-within,
+.video-timeline.is-scrubbing {
+    opacity: 1;
+    pointer-events: auto;
+    transform: translateY(0);
+}
+
+.video-timeline-track {
+    --video-playhead-position: 0%;
+    --video-loop-a-position: -100%;
+    --video-loop-b-position: -100%;
+    --video-loop-start-position: 0%;
+    --video-loop-end-position: 0%;
+    position: relative;
+    height: 18px;
+    min-width: 64px;
+    cursor: ew-resize;
+    touch-action: none;
+    outline: none;
+}
+
+.video-timeline-track::before {
+    content: "";
+    position: absolute;
+    left: 0;
+    right: 0;
+    top: 7px;
+    height: 4px;
+    border-radius: 999px;
+    background: rgba(148, 163, 184, 0.55);
+}
+
+.video-timeline-track:focus-visible::before {
+    box-shadow: 0 0 0 3px rgba(20, 184, 166, 0.42);
+}
+
+.video-timeline-buffer,
+.video-timeline-progress,
+.video-loop-range {
+    position: absolute;
+    left: 0;
+    top: 7px;
+    height: 4px;
+    border-radius: 999px;
+    pointer-events: none;
+}
+
+.video-timeline-buffer {
+    right: 0;
+    background: rgba(226, 232, 240, 0.14);
+}
+
+.video-timeline-progress {
+    width: var(--video-playhead-position);
+    background: #2dd4bf;
+    will-change: width;
+    z-index: 2;
+}
+
+.video-loop-range {
+    left: var(--video-loop-start-position);
+    right: calc(100% - var(--video-loop-end-position));
+    background: rgba(251, 191, 36, 0.32);
+    box-shadow: inset 0 0 0 1px rgba(251, 191, 36, 0.35);
+    z-index: 1;
+}
+
+.video-loop-range.is-enabled {
+    background: rgba(251, 191, 36, 0.48);
+    box-shadow:
+        inset 0 0 0 1px rgba(251, 191, 36, 0.75),
+        0 0 12px rgba(251, 191, 36, 0.2);
+}
+
+.video-timeline-thumb {
+    position: absolute;
+    left: var(--video-playhead-position);
+    top: 50%;
+    width: 12px;
+    height: 12px;
+    border-radius: 50%;
+    background: #f8fafc;
+    box-shadow: 0 1px 8px rgba(0, 0, 0, 0.45);
+    pointer-events: none;
+    transform: translate(-50%, -50%);
+    will-change: left;
+    z-index: 4;
+}
+
+.video-loop-marker {
+    position: absolute;
+    top: -4px;
+    width: 14px;
+    height: 14px;
+    border-radius: 4px;
+    background: #fbbf24;
+    color: #0f172a;
+    font-size: 9px;
+    font-weight: 900;
+    line-height: 14px;
+    text-align: center;
+    pointer-events: none;
+    transform: translateX(-50%);
+    z-index: 3;
+}
+
+.video-loop-marker-a {
+    left: var(--video-loop-a-position);
+}
+.video-loop-marker-b {
+    left: var(--video-loop-b-position);
+}
+
+.video-timeline-time {
+    min-width: 80px;
+    color: rgba(226, 232, 240, 0.92);
+    font-size: 12px;
+    font-variant-numeric: tabular-nums;
+    font-weight: 700;
+    line-height: 1;
+    text-align: right;
+    white-space: nowrap;
+}
+
+.video-loop-controls {
+    display: flex;
+    align-items: center;
+    gap: 4px;
+}
+
+.video-playback-btn,
+.video-loop-btn {
+    min-width: 24px;
+    height: 22px;
+    padding: 0 7px;
+    border: 1px solid rgba(226, 232, 240, 0.2);
+    border-radius: 6px;
+    background: rgba(15, 23, 42, 0.6);
+    color: rgba(248, 250, 252, 0.92);
+    font: inherit;
+    font-size: 11px;
+    font-weight: 800;
+    line-height: 1;
+    cursor: pointer;
+}
+
+.video-playback-btn {
+    min-width: 46px;
+}
+
+.video-playback-btn:hover,
+.video-playback-btn[aria-pressed="true"] {
+    border-color: rgba(45, 212, 191, 0.72);
+    background: rgba(45, 212, 191, 0.18);
+    color: #ccfbf1;
+}
+
+.video-loop-btn:hover:not(:disabled),
+.video-loop-btn[aria-pressed="true"] {
+    border-color: rgba(251, 191, 36, 0.72);
+    background: rgba(251, 191, 36, 0.2);
+    color: #fef3c7;
+}
+
+.video-loop-btn:disabled {
+    cursor: default;
+    opacity: 0.45;
+}
+
+.video-playback-error {
+    position: absolute;
+    left: 16px;
+    right: 16px;
+    bottom: 16px;
+    padding: 10px 12px;
+    border: 1px solid rgba(248, 113, 113, 0.55);
+    border-radius: 8px;
+    background: rgba(15, 23, 42, 0.82);
+    color: rgba(254, 226, 226, 0.96);
+    font-size: 14px;
+    font-weight: 700;
+    z-index: 2;
+}
+
+.video-lod-icon {
+    width: min(42%, 180px);
+    aspect-ratio: 1.35;
+    border-radius: 8px;
+    border: 6px solid rgba(226, 232, 240, 0.75);
+    opacity: 0.85;
+    position: relative;
+}
+
+.video-lod-icon::after {
+    content: "";
+    position: absolute;
+    left: 42%;
+    top: 26%;
+    width: 0;
+    height: 0;
+    border-top: 18px solid transparent;
+    border-bottom: 18px solid transparent;
+    border-left: 28px solid rgba(226, 232, 240, 0.8);
+}
